requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
tensorflow
numpy
opencv-python
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
sqlalchemy
alembic
psycopg2-binary
redis
celery
pytest
black
flake8
gunicorn 